Splitted statuses across files
