+++ app/models/status/aborted.rb
@@ -1,0 +1,7 @@
+module Status::Aborted
+  extend Status::Base
+
+  def self.group
+    Status::Failed
+  end
+end
