+++ app/models/status/errored.rb
@@ -1,0 +1,7 @@
+module Status::Errored
+  extend Status::Base
+
+  def self.group
+    Status::Failed
+  end
+end
