+++ app/models/status/failed.rb
@@ -1,0 +1,7 @@
+module Status::Failed
+  extend Status::Base
+
+  def self.iconize
+    {class: :danger, type: :times}
+  end
+end
