+++ app/models/status/passed_with_warnings.rb
@@ -1,0 +1,11 @@
+module Status::PassedWithWarnings
+  extend Status::Base
+
+  def passed?
+    true
+  end
+
+  def self.iconize
+    {class: :warning, type: :exclamation}
+  end
+end
